separate the function definition for interface into multiple files
