+++ pkg/devfile/parser/data/v2/commands.go
@@ -1,0 +1,49 @@
+package v2
+
+import (
+	"strings"
+
+	v1 "github.com/devfile/api/pkg/apis/workspaces/v1alpha2"
+	"github.com/devfile/parser/pkg/devfile/parser/data/common"
+)
+
+// GetCommands returns the slice of Command objects parsed from the Devfile
+func (d *DevfileV2) GetCommands() map[string]v1.Command {
+
+	commands := make(map[string]v1.Command, len(d.Commands))
+
+	for _, command := range d.Commands {
+		// we convert devfile command id to lowercase so that we can handle
+		// cases efficiently without being error prone
+		// we also convert the odo push commands from build-command and run-command flags
+		commands[common.SetIDToLower(&command)] = command
+	}
+
+	return commands
+}
+
+// AddCommands adds the slice of Command objects to the Devfile's commands
+// if a command is already defined, error out
+func (d *DevfileV2) AddCommands(commands ...v1.Command) error {
+	commandsMap := d.GetCommands()
+
+	for _, command := range commands {
+		id := common.GetID(command)
+		if _, ok := commandsMap[id]; !ok {
+			d.Commands = append(d.Commands, command)
+		} else {
+			return &common.AlreadyExistError{Name: id, Field: "command"}
+		}
+	}
+	return nil
+}
+
+// UpdateCommand updates the command with the given id
+func (d *DevfileV2) UpdateCommand(command v1.Command) {
+	id := strings.ToLower(common.GetID(command))
+	for i := range d.Commands {
+		if common.SetIDToLower(&d.Commands[i]) == id {
+			d.Commands[i] = command
+		}
+	}
+}
